Clarify process in README, make Vagrant easier to use [CON-3306]

+++ scripts/provision.sh
@@ -1,42 +1,48 @@
 #!/bin/bash
-set -ex
+set -e
 
-CONJUR_CLI_VERSION=${CONJUR_CLI_VERSION-'5.2.0'}
+CONJUR_CLI_VERSION=${CONJUR_CLI_VERSION-'5.2.5'}
 APPLIANCE_IMAGE_TAG=${APPLIANCE_IMAGE_TAG-latest}
 DOCKER_VERSION='1.9.1-0~trusty'
 
-# Install Docker
-apt-key adv \
---keyserver hkp://p80.pool.sks-keyservers.net:80 \
---recv-keys 58118E89F3A912897C070ADBF76221572C52609D
-bash -c 'echo "deb https://apt.dockerproject.org/repo ubuntu-trusty main" > /etc/apt/sources.list.d/docker.list'
-apt-get update
-apt-get install -y linux-image-extra-$(uname -r)
-apt-get install -y docker-engine=$DOCKER_VERSION
-docker run --rm hello-world
+BOXUSER=${1-'ubuntu'}
 
-# Enable the 'ubuntu' user to manage docker without sudo
-usermod -a -G docker ubuntu
+if ! docker info > /dev/null 2>&1; then
+  # Install Docker
+  apt-key adv \
+  --keyserver hkp://p80.pool.sks-keyservers.net:80 \
+  --recv-keys 58118E89F3A912897C070ADBF76221572C52609D
+  bash -c 'echo "deb https://apt.dockerproject.org/repo ubuntu-trusty main" > /etc/apt/sources.list.d/docker.list'
+  apt-get update
+  apt-get install -y linux-image-extra-$(uname -r)
+  apt-get install -y docker-engine=$DOCKER_VERSION
+
+  # Enable the 'ubuntu' or 'vagrant' user to manage docker without sudo
+  usermod -a -G docker $BOXUSER
+fi
 
 # Install fail2ban to stop malicious attackers
 apt-get install -y fail2ban ntp
 
-cli_dlpath='/tmp/conjur.deb'
-wget -q \
-"https://s3.amazonaws.com/conjur-releases/omnibus/conjur_${CONJUR_CLI_VERSION}-1_amd64.deb" -O ${cli_dlpath} && \
-dpkg -i ${cli_dlpath} && \
-rm -f ${cli_dlpath}
+if ! conjur > /dev/null 2>&1; then
+  cli_dlpath='/tmp/conjur.deb'
+  wget -q \
+    "https://github.com/conjurinc/cli-ruby/releases/download/v$CONJUR_CLI_VERSION/conjur_$CONJUR_CLI_VERSION-1_amd64.deb" \
+    -O ${cli_dlpath} && \
+    dpkg -i ${cli_dlpath} && \
+    rm -f ${cli_dlpath}
+fi
 
-echo "Loading Conjur appliance image into Docker"
-gzip -d /tmp/conjur-appliance.tar.gz # Unzip the tar.gz into a tar
-docker load < '/tmp/conjur-appliance.tar'
+if ! docker images | grep conjur; then
+  echo "Loading Conjur appliance image into Docker"
+  gzip -df /tmp/conjur-appliance.tar.gz # Unzip the tar.gz into a tar
+  docker load -i /tmp/conjur-appliance.tar
+fi
 
+image_id=$(docker images -q)
 container_name='conjur-appliance'
 
-docker rm -f ${container_name} || true  # Try to remove the container, even if it doesn't exist
-
-# Pull the Conjur UI, for caching
-docker pull conjurinc/conjur-ui
+docker rm -f ${container_name} || true
 
 # Start the appliance
 cid=$(docker create \
@@ -46,7 +52,10 @@ cid=$(docker create \
 -v /var/log/conjur:/var/log/conjur \
 -v /opt/conjur/backup:/opt/conjur/backup \
 -p "443:443" -p "636:636" -p "5432:5432" -p "5433:5433" -p "127.0.0.1:38053:38053" \
-registry.tld/conjur-appliance:${APPLIANCE_IMAGE_TAG})
+$image_id)
+
+# Clean up Docker tars, no longer needed
+rm -f /tmp/conjur-appliance.tar*
 
 cat << CONF > /etc/init/conjur.conf
 description "Conjur server"
